seems some people install from packages, and ALSO follow install.html...!
It should be obvious to anyone, installing ANY package -- not just gitolite --
that you should not do both.  It definitely should not be gitolite's job to
educate people on this, but I guess I am feeling generous today.

(and that's despite the fact that the maroon who asked this on IRC even had an
attitude!  Sadly, he quit before I saw the snark...)