step 1: figure out what works...
The genesis of this series of changes is the sudden (and very late)
discovery that github silently allows "URL/ending/in/foo" to resolve to
"URL/ending/in/foo.html" if foo.html exists and foo does not.

This means throwing the generated "site" on to a hard disk and trying to
access it either direct from the file system or using something like
`python -m http.server`, many of the links don't work!

Sure it's partly my fault for saying `[install](install)` instead of
`[install](install.mkd)`, but if I weren'tusing github I'd have realised
this long ago.

Anyway, here goes...